+++ src/main/java/com/coupang/marketplace/global/util/crypto/SaltGenerator.java
@@ -1,4 +1,4 @@
-package com.coupang.marketplace.global.util;
+package com.coupang.marketplace.global.util.crypto;
 
 import java.util.Random;
