+++ src/main/java/com/coupang/marketplace/global/util/crypto/Sha256Encryptor.java
@@ -1,4 +1,4 @@
-package com.coupang.marketplace.global.util;
+package com.coupang.marketplace.global.util.crypto;
 
 import org.springframework.stereotype.Component;
