+++ src/main/java/com/coupang/marketplace/user/service/UserService.java
@@ -3,9 +3,9 @@ package com.coupang.marketplace.user.service;
 import com.coupang.marketplace.user.controller.SignUpRequestDto;
 import com.coupang.marketplace.user.domain.User;
 import com.coupang.marketplace.user.repository.UserRepository;
-import com.coupang.marketplace.global.util.CryptoData;
-import com.coupang.marketplace.global.util.Encryptor;
-import com.coupang.marketplace.global.util.SaltGenerator;
+import com.coupang.marketplace.global.util.crypto.CryptoData;
+import com.coupang.marketplace.global.util.crypto.Encryptor;
+import com.coupang.marketplace.global.util.crypto.SaltGenerator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
